cmake: use googletest 1.10.0 from a submodule
A common use case is to integrate libraries as git submodules and a
single project might have multiple modules each with test dependencies
to gtest. Usually the top level project configures gtest, propagating
the target to it's submodules. This is preferable to the case where
each submodule uses ExternalProject to build & install gtest in a
custom location with possibly arbitrary target names (googletest,
gtest, <foobar>_gtest, etc.) as this can easily lead to multiple gtest
builds.

As a solution check if gtest (which is defined by googletest's build
files) is an existing target to determine if project's own googletest
submodule should be added to the build. Either way the subsequent
add_subdirectory() tests will have a target "gtest" to link
against. Various INTERFACE flags will take care of the rest.

A further improvement is to add "${PROJECT_NAME}::" as a prefix to the
tests. This helps when cpp_redis is a part of a larger project
and (for any reason) developers want to run cpp_redis' tests only
using "ctest -R cpp_redis".